Changed from last fm api to Spotify web api
last.fm would often not give the album in the json response, though it the album would be on their website if you searched it. Spotify web api is much more reliable, and always returns the album, though I haven't found a way to get genre, but I am fine with that.