Added main tab feature
The main tab feature allows users to queue YouTube videos in a separate tab while the main tab continues playing. When the main tab finishes, it will rotate through the queue as normal. This allows users to add videos that are not suggested to the queue. Updating the non-main tab will not affect the queue.